+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod argument;
 mod error;
+mod usage;
 pub mod parsers;
 
 pub use argument::*;
@@ -9,17 +10,18 @@ pub use parsers::literal::literal;
 pub use parsers::number::{
     float_32, float_64, integer_i16, integer_i32, integer_i64, integer_i8, integer_u16, integer_u32, integer_u64, integer_u8,
 };
+pub use usage::*;
 
 #[cfg(test)]
 mod tests {
     use std::convert::Infallible;
 
-    use crate::{integer_i32, literal, BuildExecute, Execute, Then};
+    use crate::{integer_i32, literal, BuildExecute, Execute, Then, IntoMultipleUsage, MultipleUsage, boolean};
 
     #[test]
     fn test_main() {
         let parser = literal("foo")
-            .then(integer_i32().max(10).build_exec(|i| {
+            .then(integer_i32("bar").max(10).build_exec(|i| {
                 println!("Found integer {}", i);
                 Ok::<(), Infallible>(())
             }))
@@ -30,5 +32,27 @@ mod tests {
 
         assert!(parser.execute("foo 13").is_err());
         assert_eq!(("", ()), parser.execute("foo").unwrap());
+    }
+
+    #[test]
+    fn test_usage() {
+        let parser = literal("foo")
+            .then(integer_i32("bar").max(10).build_exec(|i| {
+                println!("Found integer {}", i);
+                Ok::<(), Infallible>(())
+            }))
+            .then(boolean("buzz").build_exec(|_| Ok::<(), Infallible>(())))
+            .build_exec(|| {
+                println!("Didn't wanna give us a value aye?");
+                Ok::<(), Infallible>(())
+            });
+        let mut usage = parser.usage_gen();
+        // let usage = parser.usage_gen(None::<&'static str>);
+        let mut result = String::new();
+
+        while let Some(Ok(())) = usage.usage_next(&mut result) {
+            println!("{}", result);
+            result.clear();
+        }
     }
 }