+++ src/parsers/bool.rs
@@ -4,16 +4,20 @@ use nom::branch::alt;
 use nom::bytes::complete::tag_no_case;
 
 use super::CommandThen;
-use crate::{ArgumentMarkerDefaultImpl, CommandArgument, CommandError, Then};
+use crate::{ArgumentMarkerDefaultImpl, CommandArgument, CommandError, Then, IntoMultipleUsage, ChildUsage};
 
 /// Create a boolean parser
-pub fn boolean() -> BoolArgument { BoolArgument }
+pub fn boolean(name: &'static str) -> BoolArgument {
+    BoolArgument { name }
+}
 
 /// Boolean argument parser.
 ///
 /// This parser has no fields because it simply parses either `"true"` or
 /// `"false`".
-pub struct BoolArgument;
+pub struct BoolArgument {
+    name: &'static str,
+}
 
 impl CommandArgument<bool> for BoolArgument {
     fn parse<'a>(&self, input: &'a str) -> nom::IResult<&'a str, bool, CommandError<'a>> {
@@ -41,5 +45,21 @@ impl<E> Then<E> for BoolArgument {
             executor,
             output: PhantomData,
         }
+    }
+}
+
+impl IntoMultipleUsage for BoolArgument {
+    type Item = <[&'static str; 3] as IntoMultipleUsage>::Item;
+
+    fn usage_gen(&self) -> Self::Item {
+        self.usage_child().usage_gen()
+    }
+}
+
+impl ChildUsage for BoolArgument {
+    type Child = [&'static str; 3];
+
+    fn usage_child(&self) -> Self::Child {
+        ["<", self.name, ">"]
     }
 }