+++ src/parsers/then/argument.rs
@@ -6,7 +6,7 @@ use nom::error::{ErrorKind, FromExternalError};
 use nom::IResult;
 
 use super::ThenWrapper;
-use crate::{BuildExecute, BuildPropagate, CommandArgument, CommandError, Execute, Propagate, TaskLogic, Then};
+use crate::{BuildExecute, BuildPropagate, CommandArgument, CommandError, Execute, Propagate, TaskLogic, Then, IntoMultipleUsage, MultipleUsage, Chain, ChildUsage, Prefix, prefix};
 
 /// Default [`Then`] implementation for any argument type.
 pub struct CommandThen<A, E, O> {
@@ -43,6 +43,29 @@ where
             argument: self,
             task,
         }
+    }
+}
+
+impl<A, E, O> IntoMultipleUsage for CommandThen<A, E, O>
+where
+    A: IntoMultipleUsage + ChildUsage,
+    E: IntoMultipleUsage,
+{
+    type Item = Chain<A::Item, Prefix<(A::Child, &'static str), E::Item>>;
+
+    fn usage_gen(&self) -> Self::Item {
+        self.argument.usage_gen().chain(prefix((self.argument.usage_child(), " "), self.executor.usage_gen()))
+    }
+}
+
+impl<A, E, O> ChildUsage for CommandThen<A, E, O>
+where
+    A: ChildUsage,
+{
+    type Child = A::Child;
+
+    fn usage_child(&self) -> Self::Child {
+        self.argument.usage_child()
     }
 }
 
@@ -140,5 +163,27 @@ where
                 }
             },
         ))(input)
+    }
+}
+
+impl<A, E, C, O> IntoMultipleUsage for ThenExecutor<A, E, C, O>
+where
+    CommandThen<A, E, O>: IntoMultipleUsage,
+{
+    type Item = <CommandThen<A, E, O> as IntoMultipleUsage>::Item;
+
+    fn usage_gen(&self) -> Self::Item {
+        self.argument.usage_gen()
+    }
+}
+
+impl<A, E, C, O> ChildUsage for ThenExecutor<A, E, C, O>
+where
+    CommandThen<A, E, O>: ChildUsage,
+{
+    type Child = <CommandThen<A, E, O> as ChildUsage>::Child;
+
+    fn usage_child(&self) -> Self::Child {
+        self.argument.usage_child()
     }
 }